+++ gpxviewr/gpxviewr/urls.py
@@ -31,10 +31,12 @@ from baseweb.views import (
 
 from baseweb.viewset import (
     GPXFileViewSet,
+    GPXTrackSegmentViewSet,
 )
 
 router = DefaultRouter()
 router.register(r'gpxfile', GPXFileViewSet, basename='gpxfile')
+router.register(r'tracksegment', GPXTrackSegmentViewSet, basename='tracksegment')
 
 urlpatterns = [
     path('api/', include(router.urls)),